+++ test/writer_test.dart
@@ -1,4 +1,5 @@
 import 'dart:async';
+import 'dart:io';
 import 'dart:typed_data';
 
 import 'package:test/test.dart';
@@ -6,17 +7,25 @@ import 'package:test/test.dart';
 import 'package:tar/tar.dart' as tar;
 import 'system_tar.dart';
 
-void main() {
-  test('writes long file names', () async {
-    final name = '${'very' * 30} long name.txt';
-    final withLongName = tar.TarEntry.data(
-      tar.TarHeader(name: name, mode: 0, size: 0),
-      Uint8List(0),
-    );
+const oneMbSize = 1024 * 1024;
+const tenGbSize = oneMbSize * 1024 * 10;
 
-    final proc = await writeToTar(['--list'], Stream.value(withLongName));
-    expect(proc.lines, emits(name));
-  });
+void main() {
+  group('writes long file names', () {
+    for (final style in tar.OutputFormat.values) {
+      test(style.toString(), () async {
+        final name = '${'very' * 30} long name.txt';
+        final withLongName = tar.TarEntry.data(
+          tar.TarHeader(name: name, mode: 0, size: 0),
+          Uint8List(0),
+        );
+
+        final proc = await writeToTar(['--list'], Stream.value(withLongName),
+            format: style);
+        expect(proc.lines, emits(contains(name)));
+      });
+    }
+  }, testOn: '!windows');
 
   test('writes headers', () async {
     final date = DateTime.parse('2020-12-30 12:34');
@@ -45,12 +54,9 @@ void main() {
         ),
       ),
     );
-  });
+  }, testOn: '!windows');
 
   test('writes huge files', () async {
-    const oneMbSize = 1024 * 1024;
-    const tenGbSize = oneMbSize * 1024 * 10;
-
     final oneMb = Uint8List(oneMbSize);
     const count = tenGbSize ~/ oneMbSize;
 
@@ -65,5 +71,42 @@ void main() {
 
     final proc = await writeToTar(['--list', '--verbose'], Stream.value(entry));
     expect(proc.lines, emits(contains(tenGbSize.toString())));
+  }, testOn: '!windows');
+
+  group('refuses to write files with OutputFormat.gnu', () {
+    void shouldThrow(tar.TarEntry entry) {
+      final output = File('/dev/null').openWrite();
+      expect(
+          Stream.value(entry).pipe(
+              tar.tarWritingSink(output, format: tar.OutputFormat.gnuLongName)),
+          throwsA(isA<tar.TarException>()));
+    }
+
+    test('when they are too large', () {
+      final oneMb = Uint8List(oneMbSize);
+      const count = tenGbSize ~/ oneMbSize;
+
+      final entry = tar.TarEntry(
+        tar.TarHeader(
+          name: 'file.blob',
+          mode: 0,
+          size: tenGbSize,
+        ),
+        Stream<List<int>>.fromIterable(Iterable.generate(count, (i) => oneMb)),
+      );
+      shouldThrow(entry);
+    });
+
+    test('when they use long user names', () {
+      shouldThrow(
+        tar.TarEntry.data(
+          tar.TarHeader(
+            name: 'file.txt',
+            userName: 'this name is longer than 32 chars, which is not allowed',
+          ),
+          [],
+        ),
+      );
+    });
   });
 }